removed comments & added count in animation. optimized the usergamedata and usersettings data models and repositories. moved models into the right folder and renamed Helpers to helpers
