--- Android/app/src/main/java/com/pixelfitquest/Helpers/TypeExtensions.kt
@@ -1,8 +1,0 @@
-package com.pixelfitquest.Helpers
-
-import com.pixelfitquest.model.ExerciseType
-
-fun ExerciseType.displayName(): String = this.name
-    .replace("_", " ")
-    .lowercase()
-    .replaceFirstChar { it.uppercase() }
