+++ Android/app/src/main/java/com/pixelfitquest/PixelFitQuest.kt
@@ -4,8 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class PixelFitQuest : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
+class PixelFitQuest : Application()
