+++ Android/app/src/main/java/com/pixelfitquest/helpers/AutoSizeText.kt
