+++ Android/app/src/main/java/com/pixelfitquest/model/Slide.kt
@@ -1,0 +1,6 @@
+package com.pixelfitquest.model
+
+data class Slide(
+    val imageRes: Int,
+    val text: String
+)
