+++ Android/app/src/main/java/com/pixelfitquest/model/workout/ExerciseWithSets.kt
@@ -1,0 +1,7 @@
+package com.pixelfitquest.model.workout
+
+data class ExerciseWithSets(
+    val exercise: Exercise,
+    val sets: List<WorkoutSet>,
+    val avgWorkoutScore: Int
+)
