+++ Android/app/src/main/java/com/pixelfitquest/model/workout/WorkoutSummary.kt
@@ -1,0 +1,7 @@
+package com.pixelfitquest.model.workout
+
+data class WorkoutSummary(
+    val totalXp: Int,
+    val totalCoins: Int,
+    val avgScore: Float = 0f  // NEW: For UI display
+)
