+++ Android/app/src/main/java/com/pixelfitquest/repository/WorkoutTemplateRepository.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Query
-import com.pixelfitquest.model.WorkoutTemplate
+import com.pixelfitquest.model.workout.WorkoutTemplate
 import dagger.hilt.android.scopes.ViewModelScoped
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
