+++ Android/app/src/main/java/com/pixelfitquest/ui/view/HomeScreen.kt
@@ -46,7 +46,7 @@ import androidx.compose.ui.window.Dialog
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.pixelfitquest.R
-import com.pixelfitquest.model.Workout
+import com.pixelfitquest.model.workout.Workout
 import com.pixelfitquest.viewmodel.HomeViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
