+++ Android/app/src/main/java/com/pixelfitquest/ui/view/LoginScreen.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.pixelfitquest.helpers.SIGNUP_SCREEN
 import com.pixelfitquest.R
 import com.pixelfitquest.ext.AuthenticationButton
