+++ Android/app/src/main/java/com/pixelfitquest/ui/view/SettingsScreen.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.pixelfitquest.helpers.ExitAppCard
 import com.pixelfitquest.helpers.RemoveAccountCard
 import com.pixelfitquest.helpers.VolumeCard
