+++ Android/app/src/main/java/com/pixelfitquest/viewmodel/WorkoutViewModel.kt
@@ -3,13 +3,13 @@ package com.pixelfitquest.viewmodel
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import com.pixelfitquest.model.CharacterData
-import com.pixelfitquest.model.Exercise
-import com.pixelfitquest.model.ExerciseType
 import com.pixelfitquest.model.UserData
-import com.pixelfitquest.model.Workout
-import com.pixelfitquest.model.WorkoutFeedback
-import com.pixelfitquest.model.WorkoutPlan
-import com.pixelfitquest.model.WorkoutSet
+import com.pixelfitquest.model.enums.ExerciseType
+import com.pixelfitquest.model.enums.WorkoutFeedback
+import com.pixelfitquest.model.workout.Exercise
+import com.pixelfitquest.model.workout.Workout
+import com.pixelfitquest.model.workout.WorkoutPlan
+import com.pixelfitquest.model.workout.WorkoutSet
 import com.pixelfitquest.repository.UserRepository
 import com.pixelfitquest.repository.WorkoutRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
